+++ changelog.md
@@ -1,3 +1,1 @@
-- fix weird typing behaviour in amount edit box on ItemModifyScreen
-- fix startup on NeoForge
-- fix an error on the ModifyItemScreen, when the done button could not be pressed when outside a world
+- migrate to new `kotlin.time` API to resolve conflicts with newer `kotlinx.datetime` versions